Update Project as to restore tests success percentage to original state
After the updates below tests that were passing started to fail.
Some because of Batch 5 updates, but others because of the removal
of some of the container classes.
