+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/config/AggregateDataFlowTaskConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023-2023 the original author or authors.
+ * Copyright 2023-2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.springframework.batch.core.explore.JobExplorer;
-import org.springframework.batch.core.explore.support.JobExplorerFactoryBean;
 import org.springframework.batch.core.launch.support.SimpleJobLauncher;
 import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.batch.core.repository.support.JobRepositoryFactoryBean;
@@ -139,20 +138,6 @@ public class AggregateDataFlowTaskConfiguration {
 			factoryBean.afterPropertiesSet();
 		} catch (Throwable x) {
 			throw new RuntimeException("Exception creating JobRepository", x);
-		}
-		return factoryBean.getObject();
-	}
-
-	@Bean
-	public JobExplorer jobExplorer(DataSource dataSource, PlatformTransactionManager platformTransactionManager)
-		throws Exception {
-		JobExplorerFactoryBean factoryBean = new JobExplorerFactoryBean();
-		factoryBean.setDataSource(dataSource);
-		factoryBean.setTransactionManager(platformTransactionManager);
-		try {
-			factoryBean.afterPropertiesSet();
-		} catch (Throwable x) {
-			throw new RuntimeException("Exception creating JobExplorer", x);
 		}
 		return factoryBean.getObject();
 	}
